Update Buffer to Buffer.from Rename pollForApplications2 controller
